requirements.txt
# python==3.12.6
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.2
botocore==1.42.2
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
ddgs==9.9.2
distro==1.9.0
docx==0.2.4
dotenv==0.9.9
duckduckgo_search==8.1.1
etelemetry==0.3.1
fake-useragent==2.2.0
fastapi==0.123.4
filelock==3.20.0
fitz==0.0.1.dev2
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.1.7
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
looseversion==1.3.0
lxml==6.0.2
networkx==3.6
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
primp==0.15.0
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.6
pyparsing==3.2.5
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.5.0
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.16.0
scipy==1.16.3
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
traits==7.0.2
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0